Add the ability to populate worlds
For each space in the world, the population either generates no-one
(represented by a null value), or a character that is either alive or
the ravenous undead.

In future, hopefully this can be useful to generate different types of
people with different personalities, or different types of zombies
(Boyle zombies rather than Romero zombies, say).

I realised after coding this that it doesn't quite follow the iterator
pattern, wherein invocations of `next()` are supposed to return an
object with `done` and `value` keys. This seems like a useful thing to
add later, particularly if I want to be able to provide arrays to
iterate over in tests.