Let zombies pick which human to move toward
Given the zombies can move a single square in any direction including
diagonals (akin to a king in chess), using the Manhattan distance
(x distance + y distance) seems like as good a determinant as any. I
could have gone with something fancier to approximate the actual
distance, but this is seems fine for the moment, and I can always change
it later.