Add a `tick` to the world
This could be better. Testing the `tick` method requires testing the
world logic, the viewpoint logic and the character movement logic all
together, which doesn't fit in very nicely with the functional core
pattern, and leaves a separate integration-y test for movement. I'd like
to give this a good refactor later on, possibly extracing some kind of
Tick class; I'm not sure what the best shape for this would be just yet.
I'm hoping that adding a few more bits and pieces will make the shape
clearer.