Make zombies swarm rather than queue
Before this change, zombies would choose their optimal move, but would
stay still if that move was not available. This resulted in a line of
zombies queueing up to eat.

This change makes it so that the zombies will also consider moves
diagonal to their preferred move, which causese them to 'swarm' around
their target.

For instance, if a zombie wants to travel east towards its target, but
that is blocked, it will also consider moving north-east and south-east.
Similarly, if it wants to go north-east, it will also consider north and
east.