Factor out a Vector class
This class deals with all the bother of manual offset arithmetic. As an
added bonus, that should hopefully mean being able to change how the
zombies calculate distance, using something more akin to the Euclidean
distance; this should mean being able to remove the fix to stop the
queueing problem, as moving on a diagonal to get around an obstacle
will still reduce the Euclidean distance, even though it leave the
Manhattan distance the same.