Compare available moves by Euclidean distance
When using the Manhattan distance, if a zombie wanted to move in a
horizontal or vertical line towards its target, but that route was
blocked, there was no reason for it to take a diagonal (as moving one
space closer would be offset by moving one space sideways). This
resulted in queues of zombies, as initially fixed in 69a07a629431.

A more elegant fix is to use the Euclidean distance, in which case it
makes sense for a zombie to move diagonally around an obstacle. This
has the nice side-effect that we can remove a bunch of the hard-coded
movement options, and have the zombies compare all possible moves
by how close they can get to their preferred target. As one of these
moves is the null move (0, 0), that means the zombies will never choose
a move that takes them away from their target, as they can always not
move at all. There's also an added tweak, the distance can be used as a
tie-breaker between equally good moves, which avoids the potential for
oscillating zombies torn between two equally good (or bad) positions.

Another nice bonus: we can figure out the biting range based solely on
the `distance` property now, as anything with a distance of 2 or less
can only be up to 1 space in each direction. This reduces the interface
we expect of offsets, in that they now only need `equals`, `distance`
and `sub` -- `dx` and `dy` are implementation details.