Pull out target and obstacle offsets earlier
Once the zombie has determined whether something is a target or an
obstacle (or both), it no longer cares about the character: only the
offset. This means we can pull out the offsets into arrays at the start
and use them directly, rather than having to have extra `offset`
lookups.

This also means we can determine at the start that the zombie will not
be an obstacle to itself, simplifying the logic of whether a move is
available (because the null move will now always be unblocked).