Simplify best-target detection
Instead of having a concept of a best target and a null check, it seems
to simplify things to model the best target as being an infinite
distance away, such that no move will bring the zombie any closer. Given
the zombies are now lazy, and given equally good moves will take the one
with the shortest distance, this means we don't need a special case for
when there is no best target.

This also means we only use the `noMove` constant once (for the
biting-range exemption) and can remove it.