Pythonise all the things!
This is pretty much a direct port from the JavaScript prototype (as of
commit b5528480ed608c6debb6bac4bc6268c2f7635f12), give or take a few
Pythonly optimisations.

Why change now? In short, because I'd got what I needed out of the
JavaScript version. I wanted to delve back into the language (and, after
a few years' absence, learning some of the new language features), and
get myself re-acquainted enough with it to be comfortable digging into
more complex codebases. I think I've just about managed that now.

The other thing I was hoping to do with this project was get more
comfortable with the "Functional Core, Imperative Shell" pattern[1]. I
think I'm now at the point where writing this in JavaScript (at least in
the core language; I haven't looked into too many of the libraries and
frameworks yet) may well get in the way of me learning about this
pattern.

[1]: https://www.destroyallsoftware.com/screencasts/catalog/functional-core-imperative-shell