Improve tick logic
Previously, the CLI would wait for a set interval after each step in the
world; now, if the step logic takes a significant amount of time (for
instance, if there's a particularly large world with a lot of characters
in it), then it will only wait the amount of time it needs to make the
whole step take the right amount of time. So, if the tick is 0.5
seconds, but the logic takes 0.3 of those seconds, it will only sleep
for the remaining 2 seconds.

This has involved intorducing mocks into my tests for the first time.
Given how deeply this code is related to the side-effect of sleeping, I
couldn't think of a good way to achieve this in a functional style while
still making the tests reflect reality.