Rejig how characters choose their moves.
This removes the in-built assumption that moves are chosen based on a
single target, which opens up the possibility for humans to choose
their next move based on the nearest `n` zombies (or indeed all the
zombies they can see), which makes them less likely to run into the
clutches of the second-nearest zombie. This also pulls knowledge of the
inner details of the environment data out of the Character classes,
which should hopefully make that easier to change and/or encapsulate.