Pull out the Population class to be more agnostic
Instead of knowing that there are two types of characters (humans and
zombies), the Population class can now be presented with a list of
potential characters (or character factories), each with an associated
probability. This removes the dependency between Population and the
character classes, so it no longer makes sense to keep it in the
`character` module.