Extract a Roster class out of the World
This is a first step towards splitting up the over-responsible World
class into smaller classes with fewer responsibilities.

Roster may not be the best name for this class -- that suggests a list
of characters, rather than characters with their current positions --
but I couldn't think of a more appropriate name.

I ended up removing or tweaking some of the WorldBuilder tests, because
the Roster now enforces uniqueness of characters, which those tests
weren't all providing.