Extract an Area class to deal with bounds checks
The plan here is to abstract away some of the details of how the maths
works, so the World class doesn't have to worry about it. Ideally, it
would be nice for the World class not to have to know about explicit
width and height at all, but for the moment it still has to know about
them for its `rows` property. I may pull that out at a later date.

As part of this, I found out that one of the tests was setting up a
world with one character unintentionally out of bounds. Hooray for
checks!