Pull out some environment testing abstractions
This reduces the repetition in the tests (not necessarily a goal in
itself), and also means removing some of the `assume` statements from
the tests. In practice, an environment that has a character at the
zero vector that isn't the character under test is wrong, and
something that we'd only expect to be generated if something has
already broken.