Integrate the two action types into the world
This has involved relaxing some of the constraints in the whole-system
movement checks (the character states after a tick are no longer always
the same as before the tick, because some of them will have died).

This has also complicated the game-tick functionality beyond the point
where I can ignore the fact that it doesn't belong in the World class,
so I think that'll be next for extraction.