Simplify Hypothesis strategies with type annotations
This is only going to work on Python... 3.6 or greater, I think.

A nice feature of Hypothesis that I came across is its ability to avoid
explicit strategy construction when a class has type declarations on its
initialiser (or inherits from `typing.NamedTuple`). This means some of
the strategies that relied on manually constructing points, vectors,
areas and the like can now work out how to build these from the type
annotations. This has also helped simplify the nested construction of
bounding boxes containing the zero vector (albeit at the cost of a less
efficient search space; I'm not sure whether that'll prove a problem at
some point).