Pull movement ranges out into character state
This means the Character class doesn't need to know as much about the
details of how far characters can move, or the details of how to
construct those bounding boxes. All it needs to know is that there's a
movement range that it can iterate through.

This might be adding an extra responsibility to the BoundingBox class
that it doesn't necessarily need (in that it's now responsible for both
containment calculations and decisions about how to produce a list of
Vector objects), but I think that's ok for now. If it becomes an issue,
I can always split it apart again.