Pull the `next_action` method into Character
The idea of the world knowing what types of actions were available
never really made a whole lot of sense; now all that it needs to know
is that there's a method that returns an action (which acts as a
transformation on the roster).

This means that the Character class is now doing rather a lot, but I
think the next step will involve pulling out a common abstraction for
attacking and movement, which will then open the logic up to introduce
a "state change" action (for reanimating the dead).