Add a reanimation action to the dead
This is the last step to make this a proper zombie simulator (rather
than a weird-grey-people-murdering-everyone-else simulator). There's
still a bunch of common functionality between different actions that
I could extract, and some kind of common abstraction between the three
action types (moving, attacking, changing state), but this now counts
as a feature-complete first version of a zombie simulator.

Fixes #3.