Be explicit about how to generate limit boxes
I was originally generating any and all bounding boxes and filtering
them down to ones that contained the zero vector, but in the latest
version of Hypothesis this doesn't generate enough good test data. Since
it's possible to declare up-front which co-ordinate ranges will include
the zero vector, that'll work instead.