Pull out "nearest human" logic into `TargetVectors`
It's already responsible for figuring out what in the environment is a
human and what's a zombie, so it seems reasonable that it should know
about which one is closest.

This goes part of the way to pulling a bunch of this logic out so that
the characters can ask questions of the environment, rather than doing
all the maths themselves.