Fold `movement_strategy` back into `best_move`
The two-step process of asking the character's state what its movement
strategy is, then asking the movement strategy which move it would
like, is unnecessarily clunky and is making it harder to optimise things
in the viewpoint code (as it's passing around lower-level information
about target positions, rather than the higher-level `TargetVectors`
object).

As a nice example of this being a better approach, the handling of
corpse movement is now cleaner, as there's no incentive to re-use the
"shortest move available" strategy. Instead, we can look at the moves
available and explicitly error out if there's no option not to move.