Base character move logic on new `nearest` checks
This encapsulates a bit more of the low-level detail of vector
manipulation into the `TargetVectors` class (which knows about
characters and how to identify them) and the `Viewpoint` class (which
doesn't).

Right now, this does some fairly nasty things to the performance (and
adds a bunch more complexity than I'd like to the test suite, including
a fake `Viewpoint` so that the offsetting logic works). Hopefully,
though, all the performance problems are concentrated in the `nearest`
method on `Roster`, which should be readily optimisable with the right
data structures.