Key attacks by position, not character identity
This is another preparatory step to moving the roster so that it only
offers lookup by position, not by character identity.

Now that we've got the extra layer of abstraction, we can translate from
the target's relative position to their absolute position before passing
it to the character roster.

This has the nice benefit that we can re-use the `TargetVectors` from
our movement logic, and take advantage of the nearest-neighbour lookup
rather than looping through everyone manually in the character logic.

(For the moment, we're looping manually in the roster instead, but
that's about to change.)