Treat the Roster's internals like a dictionary
Because that's what it is, and we don't have to keep the old "list of
pairs" model around any more.

There's a slight kludge in here, that we're allowing an existing Roster
instance to pass in the set of characters it knows about, and we're
trusting that to be consistent, bypassing a bunch of validation code.
It's a bit of a hack, but it's a hack triggered by an
underscore-prefixed parameter, so it's reasonably clear it's for
internal use.

We're also not copying the positions dict on initialisation in this
branch, because we know that an existing Roster instance is never going
to modify it, because immutability.