Pass the size of the world into the Roster class
I'm planning to add some kind of recursive data structure to make
nearest-neighbour lookups faster. That's a litte easier (and has less
annoying balancing problems) if I know what size the space is that I'm
partitioning.

This is increasingly making it look like the World class has very little
to do, so I suspect that's either just going to become a `Tick` class
(to hold the simulation's turn logic) or is going to disappear entirely.