Pull out character positions into a pair of trees
Ok. So, this commit got a bit big, but hey, it's a personal project, so
I can get away with it.

The plan here was to make nearest-neighbour lookups more efficient by
sticking them into some kind of tree, rather than going through the list
of positions one by one. That turned out not to work particularly well,
because most lookups were for humans to find their nearest zombie, or
the nearest zombie to a position they were thinking of moving to. As
humans tend to cluster in groups, this meant the nearest-neighbour
lookup was having to get through rather a lot of humans before getting
to the closest zombie, which was slowing everything down. So I decided
to split the roster in two: one set of positions for the zombies, one
for everyone else.

This does unfortunately mean having to make the Roster class a little
more aware of what characters are (whereas before this they could be
just about anything where each character was unique), but all it knows
about so far is that characters have an `undead` property. That seems
like a reasonable trade-off for a boatload of improved performance.