Tweak env-based config to be more type-checkable
These were both following the same pattern:

  - Look something up from the environment
  - Check whether it's non-empty
  - If so, look it up again and do something with it

This caused grumblings from `mypy`'s type-checker, because it's
*technically* possible for `environ` to return different values on
subsequent calls (although I don't know what kind of system-level
machinations it would take to cause it to do so in this case).

It's easy enough to keep it quiet with a temporary variable, as `mypy`
is smart enough to know that checking something for truthiness rules out
`None` values, and so deals with `Optional` types.