Add type hints for the space types
Because a bunch of these types are interconnected and/or
self-referencing, there are quite a few forward-referenced types hanging
around (because the types aren't declared at the point we want to use
them).

One of the tests also needed a type-ignoring comment, because it's
testing a check that the type-checker now covers (namely, that you can't
add two Points together).