Simplify Roster and World construction
Before this commit, there were three ways to build a Roster:

- `Roster.for_value` with a mapping and an area
- `Roster.for_value` with a Roster (which would return itself)
- `Roster()` with a mapping and an area, called from the second case of
  the `for_value` class method
- `Roster()` with a whole pile of internals, which was used for building
  a new Roster by modifying an existing one without having to rebuild
  everything from a single mapping again

Meanwhile, there was one way of building a World, which would pass the
Roster-or-mapping into `Roster.for_value` and let it figure everything
out.

There's a better way, which is to have a single `__init__` method that
assigns attributes based on the type's internal structure, and however
many class methods as are necessary to translate from other
representations.

As it turns out, by replacing the single World constructor with an
initialiser that takes an Area and a Roster, and a `for_mapping` class
method that takes a width, a height and a mapping, we know the
difference between when we need to build a new Roster and when we have
an existing one to pass in. We then only need two ways of initialising a
Roster -- one from a mapping, and one based on an existing Roster --
which we can pull apart using the same pattern. The new `for_mapping`
class method then uses the logic from `_build_positions`, so that
doesn't need its own private method any more.