Add most of the type hints for characters
The actions logic is going to take a bit more wrangling; thankfully,
that doesn't need adding in just yet.

One strange thing here is that `typing.Collections[T]` doesn't seem to
work in the way I would have expected (namely, that it supports the
`__contains__` method for items of type `T`). Instead, it appears to
support the `__contains__` method for objects of any type, so I needed
to write my own version specifically for vectors.

Weird.