Pull out a smaller protocol for zombie targeting
To work around the current test code (which provides test stubs for the
`nearest_human` property), we can reduce the interface that the Undead
character state requires, so that it only needs to be able to tell where
the nearest human is.

This feels like it's going quite far down the path of interface
segregation, but that might not be too terrible an idea. Maybe it's just
that `mypy`'s type system makes that protocol a little clunkier than it
needs to be.