Add a whole bunch more types
Ok, so this one got away from me a bit, and I got into an experimenting
mood. Aside from the type hinting (and a whole pile of Protocol use for
structural typing), the main substantive change in here is that I
consolidated the `Attack` and `StateChange` actions in the `roster`
module, because they were encoding a little too much knowledge about how
characters work in this world. In practice, there are only three things
a character might want to do in this world:

- move themselves
- change themselves in some way (e.g. reanimate)
- change another character in some way (e.g. kill them)

By consolidating the two change actions into one, the knowledge about
how attacking and changing of state works can come out a level into the
`world` module. It's still not quite right, I think, but it's a step
closer.