Use `mypy`'s strict mode for non-test code
Now that there's enough type information than this is practical, let's
turn it on.

For the test code as it currently stands, this would involve an awful
lot of type faffing, especially once it got to the interaction with
Hypothesis. That's a bit further than I want to go for something I'm
doing to play around with new concepts.