Remove the "preserves characters" integration test
This test and the following one (`test_tick_keeps_zombies`) used to be a
single test that the characters always remained the same. I pulled those
apart in commit 572eddd9efaf, once I added the ability for characters to
change, but as a result this first part of the test isn't particularly
useful any more.