Make human move logic more efficient
Instead of brute-forcing all available moves and asking where the
closest zombie would be after each one, we can use computer science and
geometry to rule a whole load of potential moves out of consideration.

This speeds up the human decision-making process (which was the
bottleneck for the simulation) by rather a lot.