Remove the default argument to movement limits
In practice, this only ever got called in the test code, because there
are always edges to the world. Clean test code is all well and good, but
not when it doesn't reflect real-world usage.

(Also, by removing the extra `UnlimitedBoundingBox` class, we can rely
on a few more specifics of how `BoundingBox` works and make things more
efficient.)