Be a bit smarter about movement limits
Instead of going through all the possible moves (25 for humans, 9 for
zombies, 1 for corpses) and checking whether each move is inside the
world, we save a bunch of time by calculating the intersection of the
character's movement range and the limits of the world.

This does introduce a little more coupling between the character's
movement logic and the implementation of the world limits, but there's
already plenty of coupling from higher-level modules down into the
`space` module, so that seems fine.

Also, it's a nice performance bump not having to ask the same thing 25
times.