Change the API used to render the world
From running some benchmarking code, the simulation was spending a lot
of time running through each point in the world and figuring out what
was in it. Rather than doing that one point at a time, the world can
supply an Iterable of all the character positions (because we know we're
going to use them all every time) and the Renderer class can work out
how to translate that into text.

In the previous version of the World class, the World and Renderer APIs
were a little too similar for comfort (with the Renderer exposing lines
of text characters, and the World exposing rows of… other characters).
This feels like a neater separation, and actually gives the Renderer
some useful work to do. Plus, it's much faster.