Pull Tick logic out of the World
The World class is starting to look an awful lot like a thin proxy to
Roster, with the one exception of the tick logic. It used to be
responsible for keeping track of the size of the world, but now the
Roster needs to know about that too, so that it can handle nearest-enemy
lookups efficiently with a tree structure.

By pulling ticks out into their own class, I think the World class is
most likely going to fall out and disappear.

As a result of this, some of the old integration tests from
`test_movement.py` became tests on the `Tick` class, and some of them
weren't still useful enough to justify their existence, so I got rid of
them.