Eliminate the redundant World class
The Roster class now does everything this class does (since it needs to
know about the area of the world for efficient nearest-neighbour
lookups). As a result, there's not much need for a separate World class
any more.

This also means that Viewpoint becomes a generic type (it no longer
takes the type specificity from World, and doesn't need to care what
type of thing is nearby), but in doing so it inherits Roster's type
constraints. As such, it makes most sense to put it into the Roster
module itself, given the two are now fairly closely linked together (and
a Viewpoint is really just a Roster from a particular point of view).
Ideally we wouldn't need to have this shared type variable with its very
particular constraint of undead-ness; that's a fix for the future, I
think.

This does leave things in the slightly awkward position of having a
`world` module that doesn't really have anything to do with a world.
It's mostly logic around how to build up a new roster, and how to
progress it from one state to the next. One to fix in a later commit,
perhaps.