Refactor the two position trees into a dict
Rather than having to know about the two different class members for the
different types of characters, pulling this into a tree adds a little
extra flexibility. Also, most of the places where we use these trees
don't need to know that there are exactly two of them, as they just
iterate through all the values. Again, a touch more flexibility.

This may be the first step to pulling out some kind of partitioned tree
type, so the roster doesn't need to know about the specifics at all.