Remove the old `character_at` Viewpoint method
Now that we have a way to check for occupied points en masse
(c245df2c114f), this method is no longer used. In doing so, we get to
simplify the interface, avoid having to fix an unsightly `object` in the
type signature, and open up the option of putting barriers into the
world.