Add some extra partitioning to the roster
Rather than just partitioning by whether a character is undead or not,
we can partition by their entire state. One drawback of this is that we
lose some of the flexibility of being able to run `nearest` checks by
arbitrary attributes, but we weren't using that anyway, and it was
sidestepping `mypy`s' type-checking. For instance, there wasn't any
check that the mandatory `undead` parameter was in the set of attributes
being queried.

This also means that we're no longer ever passing in a predicate
argument, so that whole lump of functionality can disappear (and
possibly open up a performance optimisation or two) in a later commit.

This temporarily encodes more knowledge about what characters look like
into the Roster class, but opens up the possibility of passing the
partition function in at construction time, so that can migrate to a
more appropriate place. After all, all the roster needs to know about is
that the key type is hashable, and how to get hold of it.