Remove life-state dependency from roster tests
Now that the Roster class is entirely agnostic about the type of its
characters, we can replace the character-like behaviour in the tests
with something simpler. Now that we're only testing that it's possible
to move and change characters, we only need to test out one type of
change to know that any type of change should work.