Deal with character's "life states" with an Enum
The original intention of addressing this through multiple Booleans was
to have a very duck-typed implementation: to ask whether a given
character was living, or whether it was hungering for the taste of
flesh. In practice, though, this ended up being used like a three-state
enumeration, whether the fourth (True, True) combination was implicitly
disallowed.

Honestly, it's neater to have a simplified Enum type, which still isn't
aware of the internal details of how a character's state is handled, but
makes that illegal fourth state impossible to express.

Also, now that the LifeState enum lives in a module alongside the
Character class, it doesn't need an extra layer of indirection (the
HasLifeState protocol) to translate between the modules. If characters
end up splitting into multiple classes, any changes are going to be in
the same module anyway.