Move and tidy up random barrier generation
It makes sense to stick this into the `barriers` module, together with
the Barriers class itself. By making it take an Area object, we can take
out some of the assumptions about the world's co-ordinate system
starting at (0, 0).

By applying this to the Builder class as well, we can now remove some of
the grid-iteration code from that class too, now that it's built into
Area.