Prototype direction barrier rendering
Rendering the barrier points differently depending on what's next to
them makes them a lot more legible, as they connect up with each other
and form proper walls.

This still needs a whole bunch of testing (because I really don't want
to have to change this stuff without some confidence it's working), but
it proves the concept is useful.