Make the world-buildin test a bit more useful
By generating the population as a list of just the right size, we can
use it to make assertions on the builder's behaviour. Also, we can
replace the manual co-ordinate checks with `in world_area`. This does
technically introduce coupling with the `Area.__contains__` method, but
that's pretty stable at this point, and much lower-level than the code
under test here.