Add the option to leave out the empty-space dots
When the world takes up the whole terminal, and there are barriers to
break up the space, these dots get in the way more than they help. So
let's get rid of them in those situations.

The `get_world_size` method is a bit gnarly at this point (returning a
tuple of a tuple and a Boolean), but nothing too much depends on it, so
it should be something I can fix later if it changes.