Use Hypothesis' data drawing in a barrier test
Because of the complexity of this particular example (generating an area
and a point in that area, and then a bunch of areas that *don't* contain
that point), this test was getting rather unpleasant, and wasn't
producing enough good examples to get past the `assume`. Because of how
interwoven the different parameters are, data-drawing seems like the
best option.