Pre-calculate the set of barrier points
Rather than going through the area each time, it's much quicker for the
purposes of rendering to build up a set of points and iterate through
them directly. This removes the rather grotty nested loop in the
`occupied` method, which is probably where most of the performance boost
is coming from.

This doesn't do an awful lot for listing out the occupied points in a
given area, which is taking up a chunk of the current move logic. That
might require co-opting the SpaceTree implementation we're currently
using in the roster.