Refactor out the node-splitting logic
This separates the mechanics of deciding where to split a node from the
mechanics of putting the data structures together and assigning points
to child nodes.

Because this is all a little more compact now, and related bits of code
are sitting next to each other, I think it's ok to inline the midpoint
functions. (Especially as the fact that they're midpoints is incidental
to their function.)