Remove the `from_offset` viewpoint indirection
This is partly useful as prep work for potentially introducing a vision
limit for characters: because the character's position is always the
viewpoint's frame of reference, it should be easier to encode a
viewpoint with a maximum range.

Also, it ultimately makes the code neater and reduces the size of the
`Viewpoint` interface. Since the dominant use case is for humans to work
out where their nearest zombie *would* be after a move, it's reasonable
to optimise for that rather than for looking at what the nearest human
is right now for zombies.