Set up a GitHub Actions workflow to run tests
There are a whole load of ways this could be optimised, but let's see
whether it even works first.

In my local development environment, the virtual environment is set up
by `direnv`, so all the dependencies are automatically available within
a shell. In the CI environment, this environment gets set up elsewhere,
so these commands need routing through Pipenv.