Add Black-based lint checking to the build
I'm already running this a a pre-commit hook, but it's a nice
consistency check to have it in the build too.

This also involved adding Black as a development dependency, which pins
it to a specific version. That would be particularly useful if more than
one person worked on this project.